Move genericQueryCache to dedicated file
